refactor: seperate FileUploadButton
