+++ front/src/UI/atoms/FileUploadButton/FileUploadButtonQueries.ts
@@ -1,0 +1,9 @@
+import { gql } from '@apollo/client';
+
+export const SINGLE_UPLOAD = gql`
+  mutation singleUpload($uploadUserAvatarInput: UploadUserAvatarInput!) {
+    singleUpload(uploadUserAvatarInput: $uploadUserAvatarInput) {
+      success
+    }
+  }
+`;
