+++ front/src/UI/molecules/ProfileLargeEdit/ProfileLargeEditQueries.ts
@@ -1,13 +1,5 @@
 import { gql } from '@apollo/client';
 
-export const SINGLE_UPLOAD = gql`
-  mutation singleUpload($uploadUserAvatarInput: UploadUserAvatarInput!) {
-    singleUpload(uploadUserAvatarInput: $uploadUserAvatarInput) {
-      success
-    }
-  }
-`;
-
 export const UPDATE_NICKNAME = gql`
   mutation ($user: UpdateUserInput!) {
     updateUser(updateUserInput: $user) {
